+++ test/extended_http_test.dart
@@ -1,4 +1,5 @@
 import 'package:extended_http/extended_http.dart';
+import 'package:extended_http/utils.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:path_provider_platform_interface/path_provider_platform_interface.dart';
 import 'fake_path_provider.dart';
@@ -60,5 +61,57 @@ void main() {
 
     expect(uri1.toString(), "http://domain1.com/api/users");
     expect(uri2.toString(), "http://domain2.com/api/posts");
+  });
+
+  test('create custom config for a specified path', () async {
+    final http = ExtendedHttp();
+
+    http.config(
+      baseURL: 'http://domain.com',
+      headers: {
+        "Content-Type": "application/json",
+      },
+    );
+
+    final uri1 = http.createURI('/api/users');
+    final uri2 = http.createURI(
+      '/api/posts',
+      overrideConfig: HttpOptionalConfig(
+        headers: {
+          "Content-Type": "application/x-www-form-urlencoded",
+        },
+      ),
+    );
+    final uri3 = http.createURI('/api/profile');
+
+    expect(uri1.toString(), "http://domain.com/api/users");
+    expect(uri2.toString(), "http://domain.com/api/posts");
+    expect(uri3.toString(), "http://domain.com/api/profile");
+
+    expect(
+      http.headers,
+      containsPair("Content-Type", "application/json"),
+    );
+    expect(
+      http.getConfig(uri1).headers,
+      containsPair("Content-Type", "application/json"),
+    );
+    expect(
+      http.getConfig(uri2).headers,
+      containsPair("Content-Type", "application/x-www-form-urlencoded"),
+    );
+    expect(
+      http.getConfig(uri3).headers,
+      containsPair("Content-Type", "application/json"),
+    );
+  });
+
+  test('check if instances are singleton', () async {
+    final instance1 = ExtendedHttp('domain1');
+    final instance2 = ExtendedHttp('domain2');
+
+    expect(instance1, isNot(same(instance2)));
+    expect(instance1, same(ExtendedHttp('domain1')));
+    expect(instance2, same(ExtendedHttp('domain2')));
   });
 }